rt ainvs: DetSched projections as primrec
This commit converts existing AInvs DetSched projections (tcb_of, sc_of
and ep_of) to be defined as primrec. It also adds ntfn_of and reply_of.
Defining them as primrec seems to work well in terms of simplification
and often results in simpler codes.

--

So far, tcb_of and sc_of in AInvs are case statements on kernel objects
defined as definitions, and they needed to be explicitly unfolded to
simplify to an option value.

Making them abbreviations is one way to automate the simplification, but
it has a drawback in that the case statement often has unfortunate
interaction with Map.empty, which is an abbreviation of its own, that
tends to produce unwantedly expanded case expressions in subgoals.

It is probably possible to automate the simplication of definitions
by writing a simplification rule for each kernel object case and add
all of such rules to the simp set, but defining these projections as
primrec seems to work just as well, and it also often results in
shorter proofs as we can skip case distinctions which would be needed
otherwise.

Signed-off-by: Miki Tanaka <miki.tanaka@data61.csiro.au>