arm_hyp aspec+haskell: call vcpu_switch in associate_vcpu_tcb
An invocation to bind a thread to a VCPU will perform associate_vcpu_tcb.
Previously, vcpu_switch was called only on a context switch, and so
it was possible to bind the current thread to a VCPU and then not switch
to that VCPU. This change will allow us to prove that the current active
VCPU is the VCPU of the current thread.

Signed-off-by: Michael McInerney <m.mcinerney@unsw.edu.au>